app/requirements.txt
numpy>=1.19.5
pandas>=1.0.5
Pillow==7.2
tensorflow-cpu==2.4.1
keras==2.4.3
flask==1.1.2
flasgger==0.9.2
